Move infrastructure classes to their own packages
