+++ JavaModules/idispatch-server/src/main/java/net/pkhapps/idispatch/adapter/rest/infrastructure/ConstraintViolationJsonExceptionMapper.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.adapter.rest;
+package net.pkhapps.idispatch.adapter.rest.infrastructure;
 
 import javax.validation.ConstraintViolationException;
 import javax.ws.rs.core.Response;
