+++ JavaModules/idispatch-server/src/main/java/net/pkhapps/idispatch/adapter/rest/infrastructure/DefaultJsonExceptionMapper.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.adapter.rest;
+package net.pkhapps.idispatch.adapter.rest.infrastructure;
 
 import javax.ws.rs.core.Response;
 import javax.ws.rs.ext.ExceptionMapper;
