+++ JavaModules/idispatch-server/src/main/java/net/pkhapps/idispatch/adapter/rest/infrastructure/WebApplicationJsonExceptionMapper.java
@@ -1,4 +1,4 @@
-package net.pkhapps.idispatch.adapter.rest;
+package net.pkhapps.idispatch.adapter.rest.infrastructure;
 
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Response;
